+++ src/components/alert/alert.pw.tsx
@@ -1,15 +1,12 @@
 import React from "react";
 import { test, expect } from "@playwright/experimental-ct-react17";
 import {
-  alertCrossIcon,
   alertTitle,
   alertSubtitle,
   alertDialog,
 } from "../../../playwright/components/alert";
 import {
   checkAccessibility,
-  checkDialogIsInDOM,
-  checkDialogIsNotInDOM,
   waitForAnimationEnd,
 } from "../../../playwright/support/helper";
 import { CHARACTERS, SIZE } from "../../../playwright/support/constants";
@@ -53,7 +50,7 @@ test.describe("should render Alert component", () => {
     });
   });
 
-  test("with close icon button that does not close dialog when escape key pressed and disableEscKey prop is true", async ({
+  test("dialog does not close when escape key pressed and disableEscKey prop is true", async ({
     mount,
     page,
   }) => {
@@ -63,35 +60,42 @@ test.describe("should render Alert component", () => {
       </AlertComponent>,
     );
 
-    await checkDialogIsInDOM(page);
+    const alert = page.getByRole("alertdialog");
+    await alert.waitFor();
+
     await page.keyboard.press("Escape");
-    await checkDialogIsInDOM(page);
+
+    await expect(alert).toBeVisible();
   });
 
-  test("with keyboard accessible close icon button which closes the dialog when enter key is pressed", async ({
+  test("dialog closes when enter key is pressed on close button", async ({
     mount,
     page,
   }) => {
     await mount(<AlertComponent title="title">Alert</AlertComponent>);
 
-    await checkDialogIsInDOM(page);
-    const cross = alertCrossIcon(page);
-    await page.keyboard.press("Tab");
-    await expect(cross).toBeFocused();
-    await cross.press("Enter");
-    await checkDialogIsNotInDOM(page);
+    const alert = page.getByRole("alertdialog");
+    await alert.waitFor();
+
+    const closeButton = page.getByRole("button", { name: "Close" });
+    await closeButton.press("Enter");
+
+    await expect(alert).toBeHidden();
   });
 
-  test("with close icon button that closes dialog when clicked", async ({
+  test("dialog closes when close button is clicked", async ({
     mount,
     page,
   }) => {
     await mount(<AlertComponent title="title">Alert</AlertComponent>);
 
-    await checkDialogIsInDOM(page);
-    const cross = alertCrossIcon(page);
-    await cross.click();
-    await checkDialogIsNotInDOM(page);
+    const alert = page.getByRole("alertdialog");
+    await alert.waitFor();
+
+    const closeButton = page.getByRole("button", { name: "Close" });
+    await closeButton.click();
+
+    await expect(alert).toBeHidden();
   });
 
   viewportHeights.forEach((height) => {
@@ -134,24 +138,6 @@ test.describe("should render Alert component", () => {
       const alertElement = alertDialog(page);
       await expect(alertElement).toHaveCSS("width", `${width}px`);
     });
-  });
-
-  test("with close icon button that calls the onCancel callback when clicked", async ({
-    mount,
-    page,
-  }) => {
-    let callbackCount = 0;
-    await mount(
-      <AlertComponent
-        onCancel={() => {
-          callbackCount += 1;
-        }}
-      />,
-    );
-
-    const cross = alertCrossIcon(page);
-    await cross.click();
-    expect(callbackCount).toBe(1);
   });
 
   // TODO: Skipped due to flaky focus behaviour. To review in FE-6428